wm: WindowState: Reset mAnimatingExit flag while relayout visible window
The timing issue happen when a closing dialog activity request to
launch itself again before the task move to background. By default
system will play an exit animation for the dialog style window, when
the activity being move to foreground again, core will cancel previous
exit animation so the window can be visible. But the window surface
shouldn't be hide and/or destroy in onExitAnimationDone, since the
window is request to be visible again.

Bug: 342644763
Test: repeatly closeing the sample activity, verify the window surface
of app window won't be hide after the relayout request.

Merged-In: Id2718b438ef9f21b0bc4182e50b59de07c445372
Change-Id: Id2718b438ef9f21b0bc4182e50b59de07c445372
Signed-off-by: Richard Raya <rdxzv.dev@gmail.com>